+++ railties/test/application/test_test.rb
@@ -24,7 +24,7 @@ module ApplicationTests
         end
       RUBY
 
-      run_test_file 'unit/foo_test.rb'
+      assert_successful_test_run 'unit/foo_test.rb'
     end
 
     test "integration test" do
@@ -49,19 +49,48 @@ module ApplicationTests
         end
       RUBY
 
-      run_test_file 'integration/posts_test.rb'
+      assert_successful_test_run 'integration/posts_test.rb'
+    end
+
+    test "enable full backtraces on test failures" do
+      app_file 'test/unit/failing_test.rb', <<-RUBY
+        require 'test_helper'
+
+        class FailingTest < ActiveSupport::TestCase
+          def test_failure
+            raise "fail"
+          end
+        end
+      RUBY
+
+      output = run_test_file('unit/failing_test.rb', env: { "BACKTRACE" => "1" })
+      assert_match %r{/app/test/unit/failing_test\.rb}, output
     end
 
     private
-      def run_test_file(name)
-        result = ruby '-Itest', "#{app_path}/test/#{name}"
+      def assert_successful_test_run(name)
+        result = run_test_file(name)
         assert_equal 0, $?.to_i, result
       end
 
+      def run_test_file(name, options = {})
+        ruby '-Itest', "#{app_path}/test/#{name}", options
+      end
+
       def ruby(*args)
+        options = args.extract_options!
+        env = options.fetch(:env, {})
+        env["RUBYLIB"] = $:.join(':')
+
         Dir.chdir(app_path) do
-          `RUBYLIB='#{$:.join(':')}' #{Gem.ruby} #{args.join(' ')}`
+          `#{env_string(env)} #{Gem.ruby} #{args.join(' ')}`
         end
+      end
+
+      def env_string(variables)
+        variables.map do |key, value|
+          "#{key}='#{value}'"
+        end.join " "
       end
   end
 end